[CMAPT-4951] DCU - Automated Certificate Renewal - Automated Renewal of Certificates (#21)
* [CMAPT-4951] DCU - Automated Certificate Renewal - Automated Renewal of Certificates
Adding the option for command line arg of 'certs_renewal'. This allows a csv str of cert common names to be provided. It will then validate they are certs in our mapping file, renew them, update associated kube secrets, then send a Slack message containing a list of certs that were renewed, a list of pods for each kube env that that need to be rolled, and any salt updates needed.
Sample Slack message can be found in #queue_bot_test
Added unit tests for updates.

* Removing convert_csv_to_list and adding a simpler string split in certificates_renewal.
Corrected an error that was causing kube secrets not to get added to their associated secrets list for printing to Slack.
Renewal script was run for "dcu.zeus.int.dev-godaddy.com". It successufully renewed the cert and updated the associated secret. The Slack message was created in queue_bot_test but did not include the pod to be rolled - this is the error that is listed as corrected above.

* Updating Slack channel to #customer_protection after testing